move smaller functionality units to misc.py
